// web/src/api/models/RuntimeSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Runtime summary for list views.
 */
export type RuntimeSummary = {
    created: string;
    description?: string | null;
    id: number;
    name: string;
    pack_ref?: string | null;
    ref: string;
    updated: string;
};